+++ crates/wdl-lsp/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+<!-- TODO: PR number -->
+* Added support for the `textDocument/foldingRange` request ([#]()).
+
 ## 0.19.0 - 2026-04-22
 
 ## 0.18.0 - 2026-04-02